feat(settings): implement server control panel and enhanced service management
This commit introduces a comprehensive server control panel for administrators and refactors the server management UI to support detailed statistics and connection handling for Jellyfin, Jellyseerr, and Audiobookshelf.

### Key Changes:

*   **Administration & Control Panel**:
    *   Added `ControlPanelView.kt` and `ControlPanelViewModel.kt` to allow administrators to monitor active sessions, manage scheduled tasks (start/stop), and perform server power actions (restart/shutdown).
    *   Implemented real-time polling for server tasks and active sessions.
    *   Added visual indicators for "Now Playing" items in the control panel with progress tracking.
    *   Integrated WebSocket-aware snackbars to notify users when a server is restarting or has shut down.

*   **Server Management UI Refactoring**:
    *   Decomposed the monolithic server management screen into modular components: `ServerCard.kt`, `ServerDetailDialog.kt`, and service-specific tabs (`JellyfinTab.kt`, `JellyseerrTab.kt`, `AudiobookshelfTab.kt`).
    *   Enhanced `ServerCard` with dynamic "LED" indicators for active services (TMDB, MDBList) and connection types (Local, Tailscale, Remote).
    *   Introduced a segmented tab interface in the server detail dialog for better organization of service-specific stats and network settings.

*   **Data & Domain**:
    *   Updated the `User` model and database schema (v43) to include an `isAdmin` flag.
    *   Expanded `JellyfinRepository` to include administrative APIs for sessions, system power, and scheduled tasks.
    *   Updated `SessionManager` to proactively refresh admin status from the server policy.
    *   Improved `JellyfinWebSocketManager` to trigger home data refreshes upon library changes and handle server state transitions.

*   **UI/UX Improvements**:
    *   Added specialized formatters for listening duration, task execution times, and media ticks.
    *   Introduced new vector assets for admin settings, power controls, and device status.
    *   Implemented "Manage Addresses" views for each service to allow adding/deleting secondary connection URLs.
