+++ app/src/main/java/com/makd/afinity/data/database/AfinityDatabase.kt
@@ -101,7 +101,7 @@ import com.makd.afinity.data.models.user.User
             JellyfinStatsCacheEntity::class,
             AbsDownloadEntity::class,
         ],
-    version = 42,
+    version = 43,
     exportSchema = false,
 )
 @TypeConverters(AfinityTypeConverters::class)
