+++ app/src/main/java/com/makd/afinity/data/database/DatabaseMigrations.kt
@@ -1048,6 +1048,13 @@ object DatabaseMigrations {
             }
         }
 
+    val MIGRATION_42_43 =
+        object : Migration(42, 43) {
+            override fun migrate(db: SupportSQLiteDatabase) {
+                db.execSQL("ALTER TABLE users ADD COLUMN isAdmin INTEGER NOT NULL DEFAULT 0")
+            }
+        }
+
     val ALL_MIGRATIONS =
         arrayOf(
             MIGRATION_1_2,
@@ -1091,5 +1098,6 @@ object DatabaseMigrations {
             MIGRATION_39_40,
             MIGRATION_40_41,
             MIGRATION_41_42,
+            MIGRATION_42_43,
         )
 }
