+++ app/src/main/java/com/makd/afinity/ui/main/MainViewModel.kt
@@ -3,6 +3,7 @@ package com.makd.afinity.ui.main
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.makd.afinity.data.repository.AppDataRepository
+import com.makd.afinity.data.websocket.JellyfinWebSocketManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
@@ -11,8 +12,12 @@ import kotlinx.coroutines.flow.stateIn
 import javax.inject.Inject
 
 @HiltViewModel
-class MainViewModel @Inject constructor(private val appDataRepository: AppDataRepository) :
-    ViewModel() {
+class MainViewModel @Inject constructor(
+    private val appDataRepository: AppDataRepository,
+    private val webSocketManager: JellyfinWebSocketManager,
+) : ViewModel() {
+
+    val webSocketState = webSocketManager.connectionState
 
     val uiState: StateFlow<MainUiState> =
         combine(appDataRepository.userName, appDataRepository.userProfileImageUrl) { name, imageUrl
