+++ app/src/main/java/com/makd/afinity/ui/settings/servers/ServerColors.kt
@@ -1,0 +1,11 @@
+package com.makd.afinity.ui.settings.servers
+
+import androidx.compose.ui.graphics.Color
+
+internal val JellyseerrColor = Color(0xFFA88AFA)
+internal val AudiobookshelfColor = Color(0xFFCD9E46)
+internal val tmdbColor = Color(0xFF3CBEC9)
+internal val mdblistColor = Color(0xFF4283C9)
+internal val LocalColor = Color(0xFF4CAF50)
+internal val TailscaleColor = Color(0xFF2196F3)
+internal val RemoteColor = Color(0xFFFF9800)
