+++ app/src/main/res/drawable/ic_admin_panel_settings.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
+      
+    <path android:fillColor="#ffffff" android:pathData="M17,17q0.625,0 1.063,-0.438T18.5,15.5q0,-0.625 -0.438,-1.063T17,14q-0.625,0 -1.063,0.438T15.5,15.5q0,0.625 0.438,1.063T17,17ZM17,20q0.775,0 1.425,-0.363t1.05,-0.962q-0.55,-0.325 -1.175,-0.5T17,18q-0.675,0 -1.3,0.175t-1.175,0.5q0.4,0.6 1.05,0.963T17,20ZM17,22q-2.075,0 -3.538,-1.463T12,17q0,-2.075 1.463,-3.538T17,12q2.075,0 3.538,1.463T22,17q0,2.075 -1.463,3.538T17,22ZM12,22q-3.475,-0.875 -5.738,-3.988T4,11.1L4,6.375q0,-0.625 0.363,-1.125t0.937,-0.725l6,-2.25q0.35,-0.125 0.7,-0.125t0.7,0.125l6,2.25q0.575,0.225 0.938,0.725T20,6.375v4.3q-0.65,-0.325 -1.463,-0.5T17,10q-2.9,0 -4.95,2.05T10,17q0,1.55 0.588,2.8t1.487,2.175q-0.025,0 -0.037,0.013T12,22Z"/>
+    
+</vector>
