+++ app/src/main/res/drawable/ic_devices.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:strokeColor="@color/white"
+        android:strokeLineCap="round"
+        android:strokeLineJoin="round"
+        android:strokeWidth="2"
+        android:pathData="M13,9a1,1 0 0,1 1,-1h6a1,1 0 0,1 1,1v10a1,1 0 0,1 -1,1h-6a1,1 0 0,1 -1,-1v-10" />
+    <path
+        android:strokeColor="@color/white"
+        android:strokeLineCap="round"
+        android:strokeLineJoin="round"
+        android:strokeWidth="2"
+        android:pathData="M18,8v-3a1,1 0 0,0 -1,-1h-13a1,1 0 0,0 -1,1v12a1,1 0 0,0 1,1h9" />
+    <path
+        android:strokeColor="@color/white"
+        android:strokeLineCap="round"
+        android:strokeLineJoin="round"
+        android:strokeWidth="2"
+        android:pathData="M16,9h2" />
+</vector>
